Revert "feat(connect): add option for multiple account select"
