+++ src/PeraWalletConnect.ts
@@ -3,6 +3,7 @@ import WalletConnect from "@walletconnect/client";
 
 import PeraWalletConnectError from "./util/PeraWalletConnectError";
 import {
+  openPeraWalletConnectModal,
   openPeraWalletRedirectModal,
   removeModalWrapperFromDOM,
   PERA_WALLET_CONNECT_MODAL_ID,
@@ -13,9 +14,9 @@ import {
   closePeraWalletSignTxnModal,
   PERA_WALLET_IFRAME_ID,
   PERA_WALLET_MODAL_CLASSNAME,
+  PeraWalletModalConfig,
   PERA_WALLET_SIGN_TXN_MODAL_ID,
-  setupPeraWalletConnectModalCloseListener,
-  generatePeraWalletConnectModalActions
+  setupPeraWalletConnectModalCloseListener
 } from "./modal/peraWalletConnectModalUtils";
 import {
   getWalletDetailsFromStorage,
@@ -32,10 +33,7 @@ import {
   formatJsonRpcRequest
 } from "./util/transaction/transactionUtils";
 import {detectBrowser, isMobile} from "./util/device/deviceUtils";
-import {
-  PeraWalletConnectMethodOptions,
-  PeraWalletConnectOptions
-} from "./util/peraWalletTypes";
+import {AlgorandChainIDs} from "./util/peraWalletTypes";
 import {generateEmbeddedWalletURL} from "./util/peraWalletUtils";
 import appTellerManager, {PeraTeller} from "./util/network/teller/appTellerManager";
 import {getPeraWebWalletURL} from "./util/peraWalletConstants";
@@ -45,6 +43,27 @@ import {
   WAIT_FOR_TAB_MAX_TRY_COUNT,
   WAIT_FOR_TAB_TRY_INTERVAL
 } from "./util/dom/domUtils";
+
+interface PeraWalletConnectOptions {
+  bridge?: string;
+  shouldShowSignTxnToast?: boolean;
+  chainId?: AlgorandChainIDs;
+}
+
+function generatePeraWalletConnectModalActions({
+  isWebWalletAvailable,
+  shouldDisplayNewBadge,
+  shouldUseSound
+}: PeraWalletModalConfig) {
+  return {
+    open: openPeraWalletConnectModal({
+      isWebWalletAvailable,
+      shouldDisplayNewBadge,
+      shouldUseSound
+    }),
+    close: () => removeModalWrapperFromDOM(PERA_WALLET_CONNECT_MODAL_ID)
+  };
+}
 
 class PeraWalletConnect {
   bridge: string;
@@ -82,8 +101,7 @@ class PeraWalletConnect {
     resolve: (accounts: string[]) => void,
     reject: (reason?: any) => void,
     webWalletURL: string,
-    chainId: number | undefined,
-    shouldSelectSingleAccount: boolean
+    chainId: number | undefined
   ) {
     const browser = detectBrowser();
     const webWalletURLs = getPeraWebWalletURL(webWalletURL);
@@ -130,8 +148,7 @@ class PeraWalletConnect {
                     type: "CONNECT",
                     data: {
                       ...getMetaInfo(),
-                      chainId,
-                      shouldSelectSingleAccount
+                      chainId
                     }
                   },
 
@@ -176,8 +193,7 @@ class PeraWalletConnect {
                           type: "CONNECT",
                           data: {
                             ...getMetaInfo(),
-                            chainId,
-                            shouldSelectSingleAccount
+                            chainId
                           }
                         },
 
@@ -276,8 +292,7 @@ class PeraWalletConnect {
                   type: "CONNECT",
                   data: {
                     ...getMetaInfo(),
-                    chainId,
-                    shouldSelectSingleAccount
+                    chainId
                   }
                 },
 
@@ -351,7 +366,7 @@ class PeraWalletConnect {
     };
   }
 
-  connect({shouldSelectSingleAccount = false}: PeraWalletConnectMethodOptions = {}) {
+  connect() {
     return new Promise<string[]>(async (resolve, reject) => {
       try {
         // check if already connected and kill session first before creating a new one.
@@ -376,8 +391,7 @@ class PeraWalletConnect {
           resolve,
           reject,
           webWalletURL,
-          this.chainId,
-          shouldSelectSingleAccount
+          this.chainId
         );
 
         if (isWebWalletAvailable) {
@@ -391,8 +405,7 @@ class PeraWalletConnect {
           qrcodeModal: generatePeraWalletConnectModalActions({
             isWebWalletAvailable,
             shouldDisplayNewBadge,
-            shouldUseSound,
-            shouldSelectSingleAccount
+            shouldUseSound
           })
         });
