+++ src/modal/mode/desktop/_pera-wallet-connect-modal-desktop-mode.scss
@@ -192,8 +192,8 @@
 .pera-wallet-connect-qr-code-wrapper {
   width: fit-content;
 
-  margin: 16px auto 0;
-  padding: 7px;
+  margin: 24px auto 0;
+  padding: 10px;
 
   box-shadow: 0px 20px 60px rgba(26, 35, 91, 0.15), 0px 4px 12px rgba(26, 35, 91, 0.05),
     0px 1px 4px rgba(26, 35, 91, 0.06);
@@ -205,7 +205,7 @@
 
   text-align: center;
 
-  margin: 16px 0 12px;
+  margin: 32px 0 12px;
 
   font-size: 13px;
   font-weight: 500;
@@ -327,6 +327,10 @@
   .pera-wallet-connect-modal-desktop-mode {
     grid-template-columns: 1fr;
     gap: 24px;
+  }
+
+  .pera-wallet-connect-qr-code-wrapper {
+    margin-top: 16px;
   }
 
   .pera-wallet-connect-modal-download-pera-view {
