+++ src/util/peraWalletTypes.ts
@@ -10,21 +10,4 @@ interface PeraWalletDetails {
   selectedAccount: string;
 }
 
-interface PeraWalletConnectMethodOptions {
-  shouldSelectSingleAccount?: boolean;
-}
-
-interface PeraWalletConnectOptions {
-  bridge?: string;
-  shouldShowSignTxnToast?: boolean;
-  chainId?: AlgorandChainIDs;
-}
-
-export type {
-  PeraWalletType,
-  PeraWalletPlatformType,
-  PeraWalletDetails,
-  AlgorandChainIDs,
-  PeraWalletConnectMethodOptions,
-  PeraWalletConnectOptions
-};
+export type {PeraWalletType, PeraWalletPlatformType, PeraWalletDetails, AlgorandChainIDs};
